Reworked random number generation.
This is in preparation for upgraded replay ability,
which is tied to upcoming netplay functionality.

The ability to randomly choose between options was added, along with specifying what those options are

The replay will analyze those options, and record which option was chosen more directly than just a number in a list.  Thus, systems which use a different ordering for their minions and heroes will still be able to use the same replay format.