Completed reworked auras.
There are two types of auras, which affect various objects.
Auras can either be tied to a minion, in which case they are no longer in
effect when the minion is removed or silenced, or tied to an event, in
which case they are in effect until the event occurs.

Auras can affect minions, cards, and weapons.  Some auras affect only a
 single minion (e.g. Blessing of Might) and these auras are tracked by the
 minion itself.  All other auras are tracked by the player.