Generated a JSON description for all minions
Not all minions are correctly described; most minions with battlecry
or combo effects will not work.

Eventually, we can replace the Python implementation of cards with
their description in JSON, but that day is not today.

All minions from Druid, Hunter and Mage are working, along with
a majority of minions from the other classes and neutral.