Refactored spells into a new class called SpellCard.  Implemented HeroCards.
Now heros get their information from a card.  This was necessary to make Jaraxxus able to be defined in
JSON.

Note that the hero cards aren't yet defined in JSON, but we're getting there.

Now all minions are properly defined in JSON, and can be interchanged with their python definitions.