Fix for events
Check so the event hasn't been removed before we execute it. This could
for exemple happen with a combo of Nightmare and Poultryizer. If
Poultryizer kicks in on the minion with Nightmare, so will Nightmare
still try to kill it's minion which is no longer in play (it has been
transformed into a Chicken).